Re-organize file structure
