+++ lib/database/filehandler.rb
@@ -12,9 +12,6 @@ class TestFileHandler
       f = File.open('testcommands.csv','r')
       f.each_line do |line|
         if line.length > 5
-          #cmd_line = ControlCommandLine.new
-          #cmd_line.parseText( line )
-          #cmd.lines << cmd_line
           params = line.split(',')
           action = params[0].to_s
           xCoord = params[1].to_i
