+++ lib/handlers/messagehandler_base.rb
