+++ lib/handlers/messagehandler_emergencystop.rb
