+++ lib/handlers/messagehandler_logs.rb
