+++ lib/handlers/messagehandler_measurements.rb
