+++ lib/handlers/messagehandler_message.rb
