+++ lib/handlers/messagehandler_schedule.rb
