+++ lib/handlers/messagehandler_schedule_cmd_line.rb
