+++ lib/handlers/messagehandler_status.rb
