--- lib/messaging.rb
@@ -1,3 +1,0 @@
-require_relative 'messaging/messaging'
-
-Messenger.current.start
