+++ spec/handlers/messagehandler_measurements_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 require './lib/status.rb'
-#require './lib/messaging/messaging.rb'
-require './lib/messaging/messagehandler.rb'
-require './lib/messaging/messaging_test.rb'
-require './lib/messaging/messagehandler_measurements.rb'
+#require './lib/messaging/messenger.rb'
+require './lib/handlers/messagehandler.rb'
+require './spec/fixtures/stub_messenger.rb'
+require './lib/handlers/messagehandler_measurements.rb'
 
 describe MessageHandlerMeasurement do
 
@@ -15,7 +15,7 @@ describe MessageHandlerMeasurement do
 
     Status.current = Status.new
 
-    messaging = MessengerTest.new
+    messaging = StubMessenger.new
     messaging.reset
 
     @handler = MessageHandlerMeasurement.new(messaging)
