+++ spec/handlers/messagehandler_schedule_spec.rb
@@ -1,14 +1,14 @@
 require 'spec_helper'
 require './lib/status.rb'
-#require './lib/messaging/messaging.rb'
+#require './lib/messaging/messenger.rb'
 #require './lib/messagehandler_base'
-require './lib/messaging/messagehandler.rb'
-require './lib/messaging/messagehandler_schedule.rb'
-require './lib/messaging/messaging_test.rb'
+require './lib/handlers/messagehandler.rb'
+require './lib/handlers/messagehandler_schedule.rb'
+require './spec/fixtures/stub_messenger.rb'
 
 #require './lib/messagehandler_schedule'
 #require './lib/messagehandler_schedule_cmd_line'
-#require './lib/messaging/messagehandler_logs.rb'
+#require './lib/handlers/messagehandler_logs.rb'
 
 describe MessageHandlerSchedule do
 
@@ -20,7 +20,7 @@ describe MessageHandlerSchedule do
 
     Status.current = Status.new
 
-    messaging = MessengerTest.new
+    messaging = StubMessenger.new
     messaging.reset
 
     @handler = MessageHandlerSchedule.new(messaging)
