--- test/messaging/credentials_test.rb
@@ -1,13 +1,0 @@
-require 'test_helper'
-require 'test/unit'
-class TestCredentials < Minitest::Test
-  def setup
-  end
-
-  def teardown
-  end
-
-  def test_create_credentials_test_test
-    assert_equal true, false
-  end
-end
