--- test/test_helper.rb
@@ -1,2 +1,0 @@
-require 'pry'
-require 'minitest/autorun'
